Finishing up the CrashWrangler. fixes: #25

+++ Peach.Core.OS.OSX/Agent/Monitors/CrashWrangler.cs
@@ -37,6 +37,8 @@ using System.ComponentModel;
 using System.Runtime.InteropServices;
 using Peach.Core.Dom;
 using Proc = System.Diagnostics.Process;
+using System.Text.RegularExpressions;
+using NLog;
 
 namespace Peach.Core.Agent.Monitors
 {
@@ -75,6 +77,8 @@ namespace Peach.Core.Agent.Monitors
 	[Parameter("CwPidFile", typeof(string), "CrashWrangler PID file (defaults to cw.pid)", false)]
 	public class CrashWrangler : Monitor
 	{
+		static NLog.Logger logger = LogManager.GetCurrentClassLogger();
+
 		protected string _command = null;
 		protected string _arguments = null;
 		protected string _startOnCall = null;
@@ -88,7 +92,7 @@ namespace Peach.Core.Agent.Monitors
 		protected Proc _procHandler = null;
 		protected Proc _procCommand = null;
 		protected bool? _detectedFault = null;
-		protected TimeSpan _totalProcessorTime = TimeSpan.Zero;
+		protected ulong _totalProcessorTime = 0;
 
 		public CrashWrangler(string name, Dictionary<string, Variant> args)
 			: base(name, args)
@@ -130,10 +134,10 @@ namespace Peach.Core.Agent.Monitors
 			if (!DetectedFault())
 				return;
 
-			string file = File.ReadAllText(_cwLogFile);
+			string log = File.ReadAllText(_cwLogFile);
+			string summary = GenerateSummary(log);
 
-			// TODO: Parse file into a simple string
-			data.Add("CrashWrangler", file);
+			data.Add("CrashWrangler", summary);
 		}
 
 		public override bool MustStop()
@@ -195,30 +199,48 @@ namespace Peach.Core.Agent.Monitors
 
 		private bool _IsIdleCpu()
 		{
-			System.Diagnostics.Debug.Assert(_procHandler != null);
+			System.Diagnostics.Debug.Assert(_procCommand != null);
 
 			var lastTime = _totalProcessorTime;
-			_totalProcessorTime = _procCommand.TotalProcessorTime;
 
-			// TODO: TotalProcessorTime is not implemented, need to use task_info()
-			return lastTime == _totalProcessorTime && false;
+			// TODO: TotalProcessorTime is not implemented, need to use proc_pidinfo()
+
+			int len = Marshal.SizeOf(typeof(proc_taskinfo));
+			IntPtr ptr = Marshal.AllocHGlobal(len);
+
+			try
+			{
+				int err = proc_pidinfo(_procCommand.Id, PROC_PIDTASKINFO, 0, ptr, len);
+				if (err != len)
+				{
+					logger.Info("CrashWrangler: Could not measure CPU times for pid {0}, assuming idle.", _procCommand.Id);
+				}
+				else
+				{
+					proc_taskinfo ti = (proc_taskinfo)Marshal.PtrToStructure(ptr, typeof(proc_taskinfo));
+					_totalProcessorTime = ti.pti_total_user + ti.pti_total_system;
+				}
+			}
+			finally
+			{
+				Marshal.FreeHGlobal(ptr);
+			}
+
+			return lastTime == _totalProcessorTime;
 		}
 
 		private bool _IsProcessRunning()
 		{
-			return _procHandler != null && !_procHandler.HasExited && !_procCommand.HasExited;
+			return _procCommand != null && !_procCommand.HasExited && !IsZombie(_procCommand.Id);
 		}
 
 		private void _StartProcess()
 		{
-			System.Diagnostics.Debug.Assert(_procHandler == null);
-			System.Diagnostics.Debug.Assert(_procCommand == null);
-
 			if (File.Exists(_cwPidFile))
 				File.Delete(_cwPidFile);
 
 			if (File.Exists(_cwLogFile))
-				File.Delete(_cwLockFile);
+				File.Delete(_cwLogFile);
 
 			if (File.Exists(_cwLockFile))
 				File.Delete(_cwLockFile);
@@ -251,7 +273,7 @@ namespace Peach.Core.Agent.Monitors
 				throw new PeachException(string.Format("CrashWrangler: Could not start handler \"{0}\" - {1}", _execHandler, err));
 			}
 
-			_totalProcessorTime = TimeSpan.Zero;
+			_totalProcessorTime = 0;
 
 			// Wait for pid file to exist, open it up and read it
 			while (!File.Exists(_cwPidFile) && !_procHandler.HasExited)
@@ -273,10 +295,11 @@ namespace Peach.Core.Agent.Monitors
 				var log = File.Exists(_cwLogFile);
 
 				// If the exit code non-zero and no log means it was unable to run the command
-				// If the exit code is 0 or there is a log, the program ran to completion
-
 				if (ret != 0 && !log)
 					throw new PeachException("CrashWrangler: Handler could not run command \"" + _command + "\"");
+
+				// If the exit code is 0 or there is a log, the program ran to completion
+				_procCommand = null;
 			}
 		}
 
@@ -290,25 +313,31 @@ namespace Peach.Core.Agent.Monitors
 				Thread.Sleep(250);
 
 			// Killing _procCommand will cause _procHandler to exit
-			System.Diagnostics.Debug.Assert(_procCommand != null);
-
-			if (!_procCommand.HasExited)
+			// _procCommand might not exist if the program ran to completion
+			// prior to opening a handle to the pid
+			if (_procCommand != null)
 			{
-				_procCommand.CloseMainWindow();
-				_procCommand.WaitForExit(500);
-
 				if (!_procCommand.HasExited)
 				{
-					try
+					_procCommand.CloseMainWindow();
+					_procCommand.WaitForExit(500);
+
+					if (!_procCommand.HasExited)
 					{
-						_procCommand.Kill();
+						try
+						{
+							_procCommand.Kill();
+						}
+						catch (InvalidOperationException)
+						{
+							// Already exited between HasEcited and Kill()
+						}
+						_procCommand.WaitForExit();
 					}
-					catch (InvalidOperationException)
-					{
-						// Already exited between HasEcited and Kill()
-					}
-					_procCommand.WaitForExit();
 				}
+
+				_procCommand.Close();
+				_procCommand = null;
 			}
 
 			if (!_procHandler.HasExited)
@@ -318,12 +347,92 @@ namespace Peach.Core.Agent.Monitors
 
 			_procHandler.Close();
 			_procHandler = null;
-			_procCommand.Close();
-			_procHandler = null;
 		}
 
-		[DllImport("libc")]
-		private static extern IntPtr strerror(int err);
+		private static string GenerateSummary(string file)
+		{
+			StringBuilder summary = new StringBuilder();
+
+			if (file.Contains(":is_exploitable=no:"))
+				summary.Append("NotExploitable");
+			else if (file.Contains(":is_exploitable=yes:"))
+				summary.Append("Exploitable");
+			else
+				summary.Append("Unknown");
+
+			if (file.Contains("exception=EXC_BAD_ACCESS:"))
+			{
+				summary.Append("_BadAccess");
+
+				if (file.Contains(":access_Type=read:"))
+					summary.Append("_Read");
+				else if (file.Contains(":access_Type=write:"))
+					summary.Append("_Write");
+				else if (file.Contains(":access_Type=exec:"))
+					summary.Append("_Exec");
+				else if (file.Contains(":access_Type=recursion:"))
+					summary.Append("_Recursion");
+				else if (file.Contains(":access_Type=unknown:"))
+					summary.Append("_Unknown");
+			}
+			else if (file.Contains("exception=EXC_BAD_INSTRUCTION:"))
+				summary.Append("_BadInstruction");
+			else if (file.Contains("exception=EXC_ARITHMETIC:"))
+				summary.Append("_Arithmetic");
+			else if (file.Contains("exception=EXC_CRASH:"))
+				summary.Append("_Crash");
+
+			Regex reTid = new Regex(@"^Crashed Thread:\s+(\d+)", RegexOptions.Multiline);
+			Match mTid = reTid.Match(file);
+			if (mTid.Success)
+			{
+				string tid = mTid.Groups[1].Value;
+
+				string strReAddr = @"^Thread " + tid + @" Crashed:.*\n((\d+.*\s(?<addr>0x[0-9,a-f,A-F]+)\s.*\n)+)";
+				Regex reAddr = new Regex(strReAddr, RegexOptions.Multiline);
+				Match mAddr = reAddr.Match(file);
+				if (mAddr.Success)
+				{
+					var captures = mAddr.Groups["addr"].Captures;
+					for (int i = captures.Count - 1; i >= 0; --i)
+						summary.Append("_" + captures[i].Value);
+				}
+			}
+
+			return summary.ToString();
+		}
+
+		bool IsZombie(int pid)
+		{
+			int[] mib = new int[] {
+				CTL_KERN,
+				KERN_PROC,
+				KERN_PROC_PID,
+				pid
+			};
+
+			int len = kinfo_proc_size;
+			IntPtr ptr = Marshal.AllocHGlobal(len);
+
+			try
+			{
+				int ret = sysctl(mib, (uint)mib.Length, ptr, ref len, IntPtr.Zero, 0);
+				if (ret != -1)
+				{
+					extern_proc kp = (extern_proc)Marshal.PtrToStructure(ptr, typeof(extern_proc));
+					return kp.p_stat == (int)p_stat.SZOMB;
+				}
+				else
+				{
+					logger.Info("CrashWrangler: Failed to query process info for pid {0}.", pid);
+					return false;
+				}
+			}
+			finally
+			{
+				Marshal.FreeHGlobal(ptr);
+			}
+		}
 
 		private static string GetLastError(int err)
 		{
@@ -332,6 +441,99 @@ namespace Peach.Core.Agent.Monitors
 			return ret;
 		}
 
+		#region P/Invoke Stuff
+		// <libproc.h>
+		[DllImport("libc")]
+		private static extern int proc_pidinfo(int pid, int flavor, ulong arg, IntPtr buffer, int buffersize);
+
+		// <sys/proc_info.h>
+		[StructLayout(LayoutKind.Sequential)]
+		struct proc_taskinfo
+		{
+			public ulong pti_virtual_size;       /* virtual memory size (bytes) */
+			public ulong pti_resident_size;      /* resident memory size (bytes) */
+			public ulong pti_total_user;         /* total time */
+			public ulong pti_total_system;
+			public ulong pti_threads_user;       /* existing threads only */
+			public ulong pti_threads_system;
+			public int pti_policy;               /* default policy for new threads */
+			public int pti_faults;               /* number of page faults */
+			public int pti_pageins;              /* number of actual pageins */
+			public int pti_cow_faults;           /* number of copy-on-write faults */
+			public int pti_messages_sent;        /* number of messages sent */
+			public int pti_messages_received;    /* number of messages received */
+			public int pti_syscalls_mach;        /* number of mach system calls */
+			public int pti_syscalls_unix;        /* number of unix system calls */
+			public int pti_csw;                  /* number of context switches */
+			public int pti_threadnum;            /* number of threads in the task */
+			public int pti_numrunning;           /* number of running threads */
+			public int pti_priority;             /* task priority*/
+		}
+
+		// <sys/proc_info.h>
+		private static int PROC_PIDTASKINFO { get { return 4; } }
+
+		// sizeof(struct kinfo_proc)
+		private static int kinfo_proc_size { get { return 648; } }
+
+		// <sys/proc.h>
+		// Only contains the interesting parts at the beginning of the struct.
+		// However, we allocate kinfo_proc_size when calling the sysctl.
+		[StructLayout(LayoutKind.Sequential)]
+		struct extern_proc
+		{
+			public int p_starttime_tv_sec;
+			public int p_starttime_tv_usec;
+			public IntPtr p_vmspace;
+			public IntPtr p_sigacts;
+			public int p_flag;
+			public char p_stat;
+			public int p_pid;
+			public int p_oppid;
+			public int p_dupfd;
+			public IntPtr user_stack;
+			public IntPtr exit_thread;
+			public int p_debugger;
+			public int sigwait;
+			public uint p_estcpu;
+			public int p_cpticks;
+			public uint p_pctcpu;
+			public IntPtr p_wchan;
+			public IntPtr p_wmesg;
+			public uint p_swtime;
+			public uint p_slptime;
+			public uint p_realtimer_it_interval_tv_sec;
+			public uint p_realtimer_it_interval_tv_usec;
+			public uint p_realtimer_it_value_tv_sec;
+			public uint p_realtimer_it_value_tv_usec;
+			public uint p_rtime_tv_sec;
+			public uint p_rtime_tv_usec;
+			public ulong p_uticks;
+			public ulong p_sticks;
+			public ulong p_iticks;
+		}
+
+		// <sys/sysctl.h>
+		private static int CTL_KERN = 1;
+		private static int KERN_PROC = 14;
+		private static int KERN_PROC_PID = 1;
+
+		// <sys/proc.h>
+		private enum p_stat
+		{
+			SIDL   = 1, // Process being created by fork.
+			SRUN   = 2, // Currently runnable.
+			SSLEEP = 3, // Sleeping on an address.
+			SSTOP  = 4, // Process debugging or suspension.
+			SZOMB  = 5, // Awiting collection by parent.
+		}
+
+		[DllImport("libc")]
+		private static extern IntPtr strerror(int err);
+
+		[DllImport("libc")]
+		private static extern int sysctl([MarshalAs(UnmanagedType.LPArray)] int[] name, uint namelen, IntPtr oldp, ref int oldlenp, IntPtr newp, int newlen);
+		#endregion
 	}
 }
